requirements.txt
spacy
unidecode
pytorch_transformers
streamlit